+++ gpdb-20160503/core/Models/CreateSupabaseProjectRequest.cs
@@ -1,0 +1,116 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Gpdb20160503.Models
+{
+    public class CreateSupabaseProjectRequest : TeaModel {
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>Pw123456</para>
+        /// </summary>
+        [NameInMap("AccountPassword")]
+        [Validation(Required=false)]
+        public string AccountPassword { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>0c593ea1-3bea-11e9-b96b-88888888****</para>
+        /// </summary>
+        [NameInMap("ClientToken")]
+        [Validation(Required=false)]
+        public string ClientToken { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>PL0</para>
+        /// </summary>
+        [NameInMap("DiskPerformanceLevel")]
+        [Validation(Required=false)]
+        public string DiskPerformanceLevel { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>saas_iot_x86_modbustcp_lqt01</para>
+        /// </summary>
+        [NameInMap("ProjectName")]
+        [Validation(Required=false)]
+        public string ProjectName { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>1C1G</para>
+        /// </summary>
+        [NameInMap("ProjectSpec")]
+        [Validation(Required=false)]
+        public string ProjectSpec { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>cn-hangzhou</para>
+        /// </summary>
+        [NameInMap("RegionId")]
+        [Validation(Required=false)]
+        public string RegionId { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>127.0.0.1</para>
+        /// </summary>
+        [NameInMap("SecurityIPList")]
+        [Validation(Required=false)]
+        public string SecurityIPList { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>2</para>
+        /// </summary>
+        [NameInMap("StorageSize")]
+        [Validation(Required=false)]
+        public long? StorageSize { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>vsw-bp1cpq8mr64paltkb****</para>
+        /// </summary>
+        [NameInMap("VSwitchId")]
+        [Validation(Required=false)]
+        public string VSwitchId { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>vpc-bp*******************</para>
+        /// </summary>
+        [NameInMap("VpcId")]
+        [Validation(Required=false)]
+        public string VpcId { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>cn-hangzhou-h</para>
+        /// </summary>
+        [NameInMap("ZoneId")]
+        [Validation(Required=false)]
+        public string ZoneId { get; set; }
+
+    }
+
+}