+++ gpdb-20160503/core/Models/GetSupabaseProjectResponseBody.cs
@@ -1,0 +1,166 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Gpdb20160503.Models
+{
+    public class GetSupabaseProjectResponseBody : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>2019-09-08T16:00:00Z</para>
+        /// </summary>
+        [NameInMap("CreateTime")]
+        [Validation(Required=false)]
+        public string CreateTime { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>xxpassword</para>
+        /// </summary>
+        [NameInMap("DashboardPassword")]
+        [Validation(Required=false)]
+        public string DashboardPassword { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>username</para>
+        /// </summary>
+        [NameInMap("DashboardUserName")]
+        [Validation(Required=false)]
+        public string DashboardUserName { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>PL0</para>
+        /// </summary>
+        [NameInMap("DiskPerformanceLevel")]
+        [Validation(Required=false)]
+        public string DiskPerformanceLevel { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>postgres</para>
+        /// </summary>
+        [NameInMap("Engine")]
+        [Validation(Required=false)]
+        public string Engine { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>15</para>
+        /// </summary>
+        [NameInMap("EngineVersion")]
+        [Validation(Required=false)]
+        public string EngineVersion { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>192.168.0.11</para>
+        /// </summary>
+        [NameInMap("PrivateConnectUrl")]
+        [Validation(Required=false)]
+        public string PrivateConnectUrl { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>sbp-545434</para>
+        /// </summary>
+        [NameInMap("ProjectId")]
+        [Validation(Required=false)]
+        public string ProjectId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>supabase_project</para>
+        /// </summary>
+        [NameInMap("ProjectName")]
+        [Validation(Required=false)]
+        public string ProjectName { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1C1G</para>
+        /// </summary>
+        [NameInMap("ProjectSpec")]
+        [Validation(Required=false)]
+        public string ProjectSpec { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>10.154.11.10</para>
+        /// </summary>
+        [NameInMap("PublicConnectUrl")]
+        [Validation(Required=false)]
+        public string PublicConnectUrl { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>cn-hangzhou</para>
+        /// </summary>
+        [NameInMap("RegionId")]
+        [Validation(Required=false)]
+        public string RegionId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>ABB39CC3-4488-4857-905D-2E4A051D****</para>
+        /// </summary>
+        [NameInMap("RequestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>127.0.0.1</para>
+        /// </summary>
+        [NameInMap("SecurityIpList")]
+        [Validation(Required=false)]
+        public string SecurityIpList { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>running</para>
+        /// </summary>
+        [NameInMap("Status")]
+        [Validation(Required=false)]
+        public string Status { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>2</para>
+        /// </summary>
+        [NameInMap("StorageSize")]
+        [Validation(Required=false)]
+        public long? StorageSize { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>vsw-bp1cpq8mr64paltkb****</para>
+        /// </summary>
+        [NameInMap("VSwitchId")]
+        [Validation(Required=false)]
+        public string VSwitchId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>vpc-bp*******************</para>
+        /// </summary>
+        [NameInMap("VpcId")]
+        [Validation(Required=false)]
+        public string VpcId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>cn-hangzhou-i</para>
+        /// </summary>
+        [NameInMap("ZoneId")]
+        [Validation(Required=false)]
+        public string ZoneId { get; set; }
+
+    }
+
+}