+++ gpdb-20160503/core/Models/ListSupabaseProjectsResponseBody.cs
@@ -1,0 +1,224 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Gpdb20160503.Models
+{
+    public class ListSupabaseProjectsResponseBody : TeaModel {
+        [NameInMap("Items")]
+        [Validation(Required=false)]
+        public List<ListSupabaseProjectsResponseBodyItems> Items { get; set; }
+        public class ListSupabaseProjectsResponseBodyItems : TeaModel {
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>2021-10-09T04:54:08Z</para>
+            /// </summary>
+            [NameInMap("CreateTime")]
+            [Validation(Required=false)]
+            public string CreateTime { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>xxpassword</para>
+            /// </summary>
+            [NameInMap("DashboardPassword")]
+            [Validation(Required=false)]
+            public string DashboardPassword { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>null</para>
+            /// </summary>
+            [NameInMap("DashboardUserName")]
+            [Validation(Required=false)]
+            public string DashboardUserName { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>PL0</para>
+            /// </summary>
+            [NameInMap("DiskPerformanceLevel")]
+            [Validation(Required=false)]
+            public string DiskPerformanceLevel { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>gpdb</para>
+            /// </summary>
+            [NameInMap("Engine")]
+            [Validation(Required=false)]
+            public string Engine { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>6.0</para>
+            /// </summary>
+            [NameInMap("EngineVersion")]
+            [Validation(Required=false)]
+            public string EngineVersion { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>Postpaid</para>
+            /// </summary>
+            [NameInMap("PayType")]
+            [Validation(Required=false)]
+            public string PayType { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>192.168.0.1</para>
+            /// </summary>
+            [NameInMap("PrivateConnectUrl")]
+            [Validation(Required=false)]
+            public string PrivateConnectUrl { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>sbp-12***</para>
+            /// </summary>
+            [NameInMap("ProjectId")]
+            [Validation(Required=false)]
+            public string ProjectId { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>supabase_project</para>
+            /// </summary>
+            [NameInMap("ProjectName")]
+            [Validation(Required=false)]
+            public string ProjectName { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>1C1G</para>
+            /// </summary>
+            [NameInMap("ProjectSpec")]
+            [Validation(Required=false)]
+            public string ProjectSpec { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>10.154.11.10</para>
+            /// </summary>
+            [NameInMap("PublicConnectUrl")]
+            [Validation(Required=false)]
+            public string PublicConnectUrl { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>cn-hangzhou</para>
+            /// </summary>
+            [NameInMap("RegionId")]
+            [Validation(Required=false)]
+            public string RegionId { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>127.0.0.1</para>
+            /// </summary>
+            [NameInMap("SecurityIPList")]
+            [Validation(Required=false)]
+            public string SecurityIPList { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>running</para>
+            /// </summary>
+            [NameInMap("Status")]
+            [Validation(Required=false)]
+            public string Status { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>2</para>
+            /// </summary>
+            [NameInMap("StorageSize")]
+            [Validation(Required=false)]
+            public long? StorageSize { get; set; }
+
+            /// <summary>
+            /// <para>vSwitch ID。</para>
+            /// 
+            /// <b>Example:</b>
+            /// <para>vsw-bp1cpq8mr64paltkb****</para>
+            /// </summary>
+            [NameInMap("VSwitchId")]
+            [Validation(Required=false)]
+            public string VSwitchId { get; set; }
+
+            /// <summary>
+            /// <para>VPC ID。</para>
+            /// 
+            /// <b>Example:</b>
+            /// <para>vpc-bp19ame5m1r3oejns****</para>
+            /// </summary>
+            [NameInMap("VpcId")]
+            [Validation(Required=false)]
+            public string VpcId { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>cn-hangzhou-h</para>
+            /// </summary>
+            [NameInMap("ZoneId")]
+            [Validation(Required=false)]
+            public string ZoneId { get; set; }
+
+        }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>50</para>
+        /// </summary>
+        [NameInMap("MaxResults")]
+        [Validation(Required=false)]
+        public int? MaxResults { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>caeba0bbb2be03f84eb48b699f0a****</para>
+        /// </summary>
+        [NameInMap("NextToken")]
+        [Validation(Required=false)]
+        public string NextToken { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1</para>
+        /// </summary>
+        [NameInMap("PageNumber")]
+        [Validation(Required=false)]
+        public int? PageNumber { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>20</para>
+        /// </summary>
+        [NameInMap("PageRecordCount")]
+        [Validation(Required=false)]
+        public int? PageRecordCount { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>ABB39CC3-4488-4857-905D-2E4A051D****</para>
+        /// </summary>
+        [NameInMap("RequestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>2</para>
+        /// </summary>
+        [NameInMap("TotalRecordCount")]
+        [Validation(Required=false)]
+        public int? TotalRecordCount { get; set; }
+
+    }
+
+}